Initial types for kvs
